+++ worker.sh
@@ -72,9 +72,12 @@ while true; do
     # ==================================================
 
     CITY=$(grep "<city>" config.xml | sed 's/.*<city>\(.*\)<\/city>.*/\1/' | tr -d '\r' | tr -d ' ')
+    LAT=$(grep "<lat>" config.xml | sed 's/.*<lat>\(.*\)<\/lat>.*/\1/' | tr -d '\r' | tr -d ' ')
+    LON=$(grep "<lon>" config.xml | sed 's/.*<lon>\(.*\)<\/lon>.*/\1/' | tr -d '\r' | tr -d ' ')
     APIKEY=$(grep "<apikey>" config.xml | sed 's/.*<apikey>\(.*\)<\/apikey>.*/\1/' | tr -d '\r' | tr -d ' ')
 
     echo "城市: $CITY" >> "$LOG"
+    echo "座標: $LAT,$LON" >> "$LOG"
 
     # ==================================================
     # 更新圖片
@@ -82,7 +85,7 @@ while true; do
 
     echo "render.py 開始時間: $(date)" >> "$LOG"
     START_TS=$(date +%s)
-    timeout -t 90 /usr/bin/python3 render.py "$CITY" "$APIKEY" >> "$LOG" 2>&1
+    timeout -t 90 /usr/bin/python3 render.py "$CITY" "$APIKEY" "$LAT" "$LON" >> "$LOG" 2>&1
     RET=$?
     END_TS=$(date +%s)
     echo "render.py 結束時間: $(date)" >> "$LOG"